+++ Sources/Internal/Render/Highlevel/RenderSystem.cpp
@@ -82,7 +82,7 @@ RenderSystem::~RenderSystem()
 
 void RenderSystem::RenderPermanent(RenderObject * renderObject)
 {
-    DVASSERT(renderObject->GetRemoveIndex() == -1);
+    DVASSERT(renderObject->GetRemoveIndex() == static_cast<uint32>(-1));
     
 	/*on add calculate valid world bbox*/	
     renderObject->Retain();
@@ -100,7 +100,7 @@ void RenderSystem::RenderPermanent(RenderObject * renderObject)
 
 void RenderSystem::RemoveFromRender(RenderObject * renderObject)
 {
-    DVASSERT(renderObject->GetRemoveIndex() != -1);
+    DVASSERT(renderObject->GetRemoveIndex() != static_cast<uint32>(-1));
     
 //	uint32 renderBatchCount = renderObject->GetRenderBatchCount();
 //	for (uint32 k = 0; k < renderBatchCount; ++k)
@@ -190,7 +190,7 @@ void RenderSystem::SetGlobalMaterial(NMaterial *material)
     SafeRelease(globalMaterial);
     globalMaterial = SafeRetain(material);
 
-    uint32 count = renderObjectArray.size();
+    uint32 count = static_cast<uint32>(renderObjectArray.size());
     for(uint32 i = 0; i < count; ++i)
     {
         RenderObject *obj = renderObjectArray[i];
@@ -232,7 +232,7 @@ void RenderSystem::RegisterForUpdate(IRenderUpdatable * updatable)
     
 void RenderSystem::UnregisterFromUpdate(IRenderUpdatable * updatable)
 {
-    uint32 size = objectsForUpdate.size();
+    uint32 size = static_cast<uint32>(objectsForUpdate.size());
 	for(uint32 i = 0; i < size; ++i)
 	{
 		if(objectsForUpdate[i] == updatable)
@@ -256,7 +256,7 @@ void RenderSystem::FindNearestLights(RenderObject * renderObject)
     float32 squareMinDistance = 10000000.0f;
     Vector3 position = renderObject->GetWorldBoundingBox().GetCenter();
     
-    uint32 size = lights.size();
+    uint32 size = static_cast<uint32>(lights.size());
 	
 	if(1 == size)
 	{
@@ -352,7 +352,7 @@ void RenderSystem::Update(float32 timeElapsed)
 		movedLights.clear();
     }
     
-	uint32 size = objectsForUpdate.size();
+	uint32 size = static_cast<uint32>(objectsForUpdate.size());
 	for(uint32 i = 0; i < size; ++i)
 	{
         objectsForUpdate[i]->RenderUpdate(mainCamera, timeElapsed);
