+++ Sources/Internal/Scene3D/Systems/AnimationSystem.cpp
@@ -62,7 +62,7 @@ void AnimationSystem::Process(float32 timeElapsed)
 {
     TIME_PROFILE("AnimationSystem::Process");
 
-    int componentsCount = activeComponents.size();
+    int componentsCount = static_cast<int32>(activeComponents.size());
     for(int i = 0; i < componentsCount; i++) 
     {
         AnimationComponent * comp = activeComponents[i];
